Move BufWriter to a separate file.
