+++ packages/yew/src/server_renderer.rs
@@ -1,93 +1,8 @@
-use std::borrow::Cow;
-
-use futures::channel::mpsc::{self, UnboundedSender};
 use futures::stream::{Stream, StreamExt};
 
 use crate::html::{BaseComponent, Scope};
+use crate::io::{BufWriter, DEFAULT_BUF_SIZE};
 use crate::platform::{run_pinned, spawn_local};
-
-// Same as std::io::BufWriter and futures::io::BufWriter.
-const DEFAULT_BUF_SIZE: usize = 8 * 1024;
-
-/// A [`futures::io::BufWriter`], but operates over string and yields into a Stream.
-pub(crate) struct BufWriter {
-    buf: String,
-    tx: UnboundedSender<String>,
-    capacity: usize,
-}
-
-// Implementation Notes:
-//
-// When jemalloc is used, performance of the BufWriter is related to the number of allocations
-// instead of the amount of memory that is allocated.
-//
-// A bytes::Bytes-based implementation is also tested, and yielded a similar performance.
-//
-// Having a String-based buffer avoids `unsafe { str::from_utf8_unchecked(..) }` or performance
-// penalty by `str::from_utf8(..)` when converting back to String when text based content is needed
-// (e.g.: post-processing).
-//
-// `Bytes::from` can be used to convert a `String` to `Bytes` if the web server asks for an
-// `impl Stream<Item = Bytes>`. This conversion incurs no memory allocation.
-impl BufWriter {
-    pub fn with_capacity(capacity: usize) -> (Self, impl Stream<Item = String>) {
-        let (tx, rx) = mpsc::unbounded::<String>();
-
-        let this = Self {
-            buf: String::with_capacity(capacity),
-            tx,
-            capacity,
-        };
-
-        (this, rx)
-    }
-
-    #[inline]
-    pub fn capacity(&self) -> usize {
-        self.capacity
-    }
-
-    fn drain(&mut self) {
-        let _ = self.tx.unbounded_send(self.buf.drain(..).collect());
-        self.buf.reserve(self.capacity);
-    }
-
-    /// Writes a string into the buffer, optionally drains the buffer.
-    pub fn write(&mut self, s: Cow<'_, str>) {
-        if self.buf.capacity() < s.len() {
-            // There isn't enough capacity, we drain the buffer.
-            self.drain();
-        }
-
-        // It's important to check self.buf.capacity() >= s.len():
-        //
-        // 1. self.buf.reserve() may choose to over reserve than capacity.
-        // 2. When self.buf.capacity() == s.len(), the previous buffer is not drained. So it needs
-        //    to push onto the buffer instead of sending.
-        if self.buf.capacity() >= s.len() {
-            // The next part is going to fit into the buffer, we push it onto the buffer.
-            self.buf.push_str(&s);
-        } else {
-            // if the next part is more than buffer size, we send the next part.
-
-            // We don't need to drain the buffer here as the self.buf.capacity() only changes if
-            // the buffer was drained. If the buffer capacity didn't change, then it means
-            // self.buf.capacity() > s.len() which will be guaranteed to be matched by
-            // self.buf.capacity() >= s.len().
-            let _ = self.tx.unbounded_send(s.into_owned());
-        }
-    }
-}
-
-impl Drop for BufWriter {
-    fn drop(&mut self) {
-        if !self.buf.is_empty() {
-            let mut buf = String::new();
-            std::mem::swap(&mut buf, &mut self.buf);
-            let _ = self.tx.unbounded_send(buf);
-        }
-    }
-}
 
 /// A Yew Server-side Renderer that renders on the current thread.
 #[cfg_attr(documenting, doc(cfg(feature = "ssr")))]
