+++ packages/yew/src/virtual_dom/vlist.rs
@@ -160,7 +160,7 @@ mod feat_ssr {
 
     use super::*;
     use crate::html::AnyScope;
-    use crate::server_renderer::BufWriter;
+    use crate::io::BufWriter;
 
     impl VList {
         pub(crate) async fn render_into_stream(
