+++ packages/yew/src/virtual_dom/vtag.rs
@@ -430,7 +430,7 @@ impl PartialEq for VTag {
 mod feat_ssr {
     use super::*;
     use crate::html::AnyScope;
-    use crate::server_renderer::BufWriter;
+    use crate::io::BufWriter;
     use crate::virtual_dom::VText;
 
     // Elements that cannot have any child elements.
